--- Calendar2/.gitignore
@@ -1,30 +1,0 @@
-# Xcode #
-# gitignore contributors: remember to update Global/Xcode.gitignore, Objective-C.gitignore & Swift.gitignore
-## Build generated build/ DerivedData/
-## Various settings *.pbxuser !default.pbxuser *.mode1v3 !default.mode1v3 *.mode2v3 !default.mode2v3 *.perspectivev3 !default.perspectivev3 xcuserdata/
-## Other *.moved-aside *.xcuserstate
-## Obj-C/Swift specific *.hmap
-*.ipa
-*.dSYM.zip
-*.dSYM
-## Playgrounds timeline.xctimeline playground.xcworkspace
-# Swift Package Manager #
-# Add this line if you want to avoid checking in source code from Swift Package Manager dependencies.
-# Packages/ .build/
-# CocoaPods #
-# We recommend against adding the Pods directory to your .gitignore. However
-# you should judge for yourself, the pros and cons are mentioned at:
-# https://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control #
-# Pods/
-# cocoapods-keys .env
-# Carthage #
-# Add this line if you want to avoid checking in source code from Carthage dependencies.
-# Carthage/Checkouts
-Carthage/Build
-# fastlane #
-# It is recommended to not store the screenshots in the git repo. Instead, use fastlane to re-generate the
-# screenshots whenever they are needed.
-# For more information about the recommended setup visit:
-# https://github.com/fastlane/fastlane/blob/master/fastlane/docs/Gitignore.md
-fastlane/report.xml fastlane/Preview.html fastlane/screenshots fastlane/test_output
-# others *.swp !.gitkeep .DS_Store
