--- Calendar2/PULL_REQUEST_TEMPLATE.md
@@ -1,4 +1,0 @@
-### 課題番号
-
-#### レビューしてほしい箇所
-
