+++ CalendarApp/CalendarApp.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:Calendar2.xcodeproj">
+      location = "self:CalendarApp.xcodeproj">
    </FileRef>
 </Workspace>
