+++ CalendarApp/CalendarApp.xcodeproj/xcuserdata/yonedahiroya.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>Calendar2.xcscheme_^#shared#^_</key>
+		<key>CalendarApp.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
