+++ CalendarApp/CalendarApp/Assets.xcassets/AppIcon.appiconset/Contents.json
