+++ CalendarApp/CalendarApp/Assets.xcassets/Contents.json
